Refactor init action input handling
-  Move to separate file
-  Have all essentials under one object; do not pass plugin name around separately

JIRA: RAIL-3428
